# Add cuda-profiler-api package