--- a/.azure-pipelines/azure-pipelines-linux.yml
+++ b/.azure-pipelines/azure-pipelines-linux.yml
@@ -8,150 +8,150 @@ jobs:
     vmImage: ubuntu-latest
   strategy:
     matrix:
-      linux_64_cuda_compilerNonecuda_compiler_versionNonepython3.10.____cpython:
-        CONFIG: linux_64_cuda_compilerNonecuda_compiler_versionNonepython3.10.____cpython
-        UPLOAD_PACKAGES: 'True'
-        DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cos7-x86_64
-      linux_64_cuda_compilerNonecuda_compiler_versionNonepython3.11.____cpython:
-        CONFIG: linux_64_cuda_compilerNonecuda_compiler_versionNonepython3.11.____cpython
-        UPLOAD_PACKAGES: 'True'
-        DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cos7-x86_64
-      linux_64_cuda_compilerNonecuda_compiler_versionNonepython3.8.____cpython:
-        CONFIG: linux_64_cuda_compilerNonecuda_compiler_versionNonepython3.8.____cpython
-        UPLOAD_PACKAGES: 'True'
-        DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cos7-x86_64
-      linux_64_cuda_compilerNonecuda_compiler_versionNonepython3.9.____cpython:
-        CONFIG: linux_64_cuda_compilerNonecuda_compiler_versionNonepython3.9.____cpython
-        UPLOAD_PACKAGES: 'True'
-        DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cos7-x86_64
-      linux_64_cuda_compilernvcccuda_compiler_version11.2python3.10.____cpython:
-        CONFIG: linux_64_cuda_compilernvcccuda_compiler_version11.2python3.10.____cpython
+      linux_64_c_compiler_version10cuda_compilernvcccuda_compiler_version11.2python3.10.____cpython:
+        CONFIG: linux_64_c_compiler_version10cuda_compilernvcccuda_compiler_version11.2python3.10.____cpython
         UPLOAD_PACKAGES: 'True'
         DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cuda:11.2
-      linux_64_cuda_compilernvcccuda_compiler_version11.2python3.11.____cpython:
-        CONFIG: linux_64_cuda_compilernvcccuda_compiler_version11.2python3.11.____cpython
+      linux_64_c_compiler_version10cuda_compilernvcccuda_compiler_version11.2python3.11.____cpython:
+        CONFIG: linux_64_c_compiler_version10cuda_compilernvcccuda_compiler_version11.2python3.11.____cpython
         UPLOAD_PACKAGES: 'True'
         DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cuda:11.2
-      linux_64_cuda_compilernvcccuda_compiler_version11.2python3.8.____cpython:
-        CONFIG: linux_64_cuda_compilernvcccuda_compiler_version11.2python3.8.____cpython
+      linux_64_c_compiler_version10cuda_compilernvcccuda_compiler_version11.2python3.8.____cpython:
+        CONFIG: linux_64_c_compiler_version10cuda_compilernvcccuda_compiler_version11.2python3.8.____cpython
         UPLOAD_PACKAGES: 'True'
         DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cuda:11.2
-      linux_64_cuda_compilernvcccuda_compiler_version11.2python3.9.____cpython:
-        CONFIG: linux_64_cuda_compilernvcccuda_compiler_version11.2python3.9.____cpython
+      linux_64_c_compiler_version10cuda_compilernvcccuda_compiler_version11.2python3.9.____cpython:
+        CONFIG: linux_64_c_compiler_version10cuda_compilernvcccuda_compiler_version11.2python3.9.____cpython
         UPLOAD_PACKAGES: 'True'
         DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cuda:11.2
-      linux_64_cuda_compilernvcccuda_compiler_version11.8python3.10.____cpython:
-        CONFIG: linux_64_cuda_compilernvcccuda_compiler_version11.8python3.10.____cpython
+      linux_64_c_compiler_version11cuda_compilernvcccuda_compiler_version11.8python3.10.____cpython:
+        CONFIG: linux_64_c_compiler_version11cuda_compilernvcccuda_compiler_version11.8python3.10.____cpython
         UPLOAD_PACKAGES: 'True'
         DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cuda:11.8
-      linux_64_cuda_compilernvcccuda_compiler_version11.8python3.11.____cpython:
-        CONFIG: linux_64_cuda_compilernvcccuda_compiler_version11.8python3.11.____cpython
+      linux_64_c_compiler_version11cuda_compilernvcccuda_compiler_version11.8python3.11.____cpython:
+        CONFIG: linux_64_c_compiler_version11cuda_compilernvcccuda_compiler_version11.8python3.11.____cpython
         UPLOAD_PACKAGES: 'True'
         DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cuda:11.8
-      linux_64_cuda_compilernvcccuda_compiler_version11.8python3.8.____cpython:
-        CONFIG: linux_64_cuda_compilernvcccuda_compiler_version11.8python3.8.____cpython
+      linux_64_c_compiler_version11cuda_compilernvcccuda_compiler_version11.8python3.8.____cpython:
+        CONFIG: linux_64_c_compiler_version11cuda_compilernvcccuda_compiler_version11.8python3.8.____cpython
         UPLOAD_PACKAGES: 'True'
         DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cuda:11.8
-      linux_64_cuda_compilernvcccuda_compiler_version11.8python3.9.____cpython:
-        CONFIG: linux_64_cuda_compilernvcccuda_compiler_version11.8python3.9.____cpython
+      linux_64_c_compiler_version11cuda_compilernvcccuda_compiler_version11.8python3.9.____cpython:
+        CONFIG: linux_64_c_compiler_version11cuda_compilernvcccuda_compiler_version11.8python3.9.____cpython
         UPLOAD_PACKAGES: 'True'
         DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cuda:11.8
-      linux_aarch64_cuda_compilerNonecuda_compiler_versionNonepython3.10.____cpython:
-        CONFIG: linux_aarch64_cuda_compilerNonecuda_compiler_versionNonepython3.10.____cpython
+      linux_64_c_compiler_version12cuda_compilerNonecuda_compiler_versionNonepython3.10.____cpython:
+        CONFIG: linux_64_c_compiler_version12cuda_compilerNonecuda_compiler_versionNonepython3.10.____cpython
         UPLOAD_PACKAGES: 'True'
         DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cos7-x86_64
-      linux_aarch64_cuda_compilerNonecuda_compiler_versionNonepython3.11.____cpython:
-        CONFIG: linux_aarch64_cuda_compilerNonecuda_compiler_versionNonepython3.11.____cpython
+      linux_64_c_compiler_version12cuda_compilerNonecuda_compiler_versionNonepython3.11.____cpython:
+        CONFIG: linux_64_c_compiler_version12cuda_compilerNonecuda_compiler_versionNonepython3.11.____cpython
         UPLOAD_PACKAGES: 'True'
         DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cos7-x86_64
-      linux_aarch64_cuda_compilerNonecuda_compiler_versionNonepython3.8.____cpython:
-        CONFIG: linux_aarch64_cuda_compilerNonecuda_compiler_versionNonepython3.8.____cpython
+      linux_64_c_compiler_version12cuda_compilerNonecuda_compiler_versionNonepython3.8.____cpython:
+        CONFIG: linux_64_c_compiler_version12cuda_compilerNonecuda_compiler_versionNonepython3.8.____cpython
         UPLOAD_PACKAGES: 'True'
         DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cos7-x86_64
-      linux_aarch64_cuda_compilerNonecuda_compiler_versionNonepython3.9.____cpython:
-        CONFIG: linux_aarch64_cuda_compilerNonecuda_compiler_versionNonepython3.9.____cpython
+      linux_64_c_compiler_version12cuda_compilerNonecuda_compiler_versionNonepython3.9.____cpython:
+        CONFIG: linux_64_c_compiler_version12cuda_compilerNonecuda_compiler_versionNonepython3.9.____cpython
         UPLOAD_PACKAGES: 'True'
         DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cos7-x86_64
-      linux_aarch64_cuda_compilernvcccuda_compiler_version11.2python3.10.____cpython:
-        CONFIG: linux_aarch64_cuda_compilernvcccuda_compiler_version11.2python3.10.____cpython
+      linux_aarch64_c_compiler_version10cuda_compilernvcccuda_compiler_version11.2python3.10.____cpython:
+        CONFIG: linux_aarch64_c_compiler_version10cuda_compilernvcccuda_compiler_version11.2python3.10.____cpython
         UPLOAD_PACKAGES: 'True'
         DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cuda:11.2
-      linux_aarch64_cuda_compilernvcccuda_compiler_version11.2python3.11.____cpython:
-        CONFIG: linux_aarch64_cuda_compilernvcccuda_compiler_version11.2python3.11.____cpython
+      linux_aarch64_c_compiler_version10cuda_compilernvcccuda_compiler_version11.2python3.11.____cpython:
+        CONFIG: linux_aarch64_c_compiler_version10cuda_compilernvcccuda_compiler_version11.2python3.11.____cpython
         UPLOAD_PACKAGES: 'True'
         DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cuda:11.2
-      linux_aarch64_cuda_compilernvcccuda_compiler_version11.2python3.8.____cpython:
-        CONFIG: linux_aarch64_cuda_compilernvcccuda_compiler_version11.2python3.8.____cpython
+      linux_aarch64_c_compiler_version10cuda_compilernvcccuda_compiler_version11.2python3.8.____cpython:
+        CONFIG: linux_aarch64_c_compiler_version10cuda_compilernvcccuda_compiler_version11.2python3.8.____cpython
         UPLOAD_PACKAGES: 'True'
         DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cuda:11.2
-      linux_aarch64_cuda_compilernvcccuda_compiler_version11.2python3.9.____cpython:
-        CONFIG: linux_aarch64_cuda_compilernvcccuda_compiler_version11.2python3.9.____cpython
+      linux_aarch64_c_compiler_version10cuda_compilernvcccuda_compiler_version11.2python3.9.____cpython:
+        CONFIG: linux_aarch64_c_compiler_version10cuda_compilernvcccuda_compiler_version11.2python3.9.____cpython
         UPLOAD_PACKAGES: 'True'
         DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cuda:11.2
-      linux_aarch64_cuda_compilernvcccuda_compiler_version11.8python3.10.____cpython:
-        CONFIG: linux_aarch64_cuda_compilernvcccuda_compiler_version11.8python3.10.____cpython
+      linux_aarch64_c_compiler_version11cuda_compilernvcccuda_compiler_version11.8python3.10.____cpython:
+        CONFIG: linux_aarch64_c_compiler_version11cuda_compilernvcccuda_compiler_version11.8python3.10.____cpython
         UPLOAD_PACKAGES: 'True'
         DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cuda:11.8
-      linux_aarch64_cuda_compilernvcccuda_compiler_version11.8python3.11.____cpython:
-        CONFIG: linux_aarch64_cuda_compilernvcccuda_compiler_version11.8python3.11.____cpython
+      linux_aarch64_c_compiler_version11cuda_compilernvcccuda_compiler_version11.8python3.11.____cpython:
+        CONFIG: linux_aarch64_c_compiler_version11cuda_compilernvcccuda_compiler_version11.8python3.11.____cpython
         UPLOAD_PACKAGES: 'True'
         DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cuda:11.8
-      linux_aarch64_cuda_compilernvcccuda_compiler_version11.8python3.8.____cpython:
-        CONFIG: linux_aarch64_cuda_compilernvcccuda_compiler_version11.8python3.8.____cpython
+      linux_aarch64_c_compiler_version11cuda_compilernvcccuda_compiler_version11.8python3.8.____cpython:
+        CONFIG: linux_aarch64_c_compiler_version11cuda_compilernvcccuda_compiler_version11.8python3.8.____cpython
         UPLOAD_PACKAGES: 'True'
         DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cuda:11.8
-      linux_aarch64_cuda_compilernvcccuda_compiler_version11.8python3.9.____cpython:
-        CONFIG: linux_aarch64_cuda_compilernvcccuda_compiler_version11.8python3.9.____cpython
+      linux_aarch64_c_compiler_version11cuda_compilernvcccuda_compiler_version11.8python3.9.____cpython:
+        CONFIG: linux_aarch64_c_compiler_version11cuda_compilernvcccuda_compiler_version11.8python3.9.____cpython
         UPLOAD_PACKAGES: 'True'
         DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cuda:11.8
-      linux_ppc64le_cuda_compilerNonecuda_compiler_versionNonepython3.10.____cpython:
-        CONFIG: linux_ppc64le_cuda_compilerNonecuda_compiler_versionNonepython3.10.____cpython
+      linux_aarch64_c_compiler_version12cuda_compilerNonecuda_compiler_versionNonepython3.10.____cpython:
+        CONFIG: linux_aarch64_c_compiler_version12cuda_compilerNonecuda_compiler_versionNonepython3.10.____cpython
         UPLOAD_PACKAGES: 'True'
         DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cos7-x86_64
-      linux_ppc64le_cuda_compilerNonecuda_compiler_versionNonepython3.11.____cpython:
-        CONFIG: linux_ppc64le_cuda_compilerNonecuda_compiler_versionNonepython3.11.____cpython
+      linux_aarch64_c_compiler_version12cuda_compilerNonecuda_compiler_versionNonepython3.11.____cpython:
+        CONFIG: linux_aarch64_c_compiler_version12cuda_compilerNonecuda_compiler_versionNonepython3.11.____cpython
         UPLOAD_PACKAGES: 'True'
         DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cos7-x86_64
-      linux_ppc64le_cuda_compilerNonecuda_compiler_versionNonepython3.8.____cpython:
-        CONFIG: linux_ppc64le_cuda_compilerNonecuda_compiler_versionNonepython3.8.____cpython
+      linux_aarch64_c_compiler_version12cuda_compilerNonecuda_compiler_versionNonepython3.8.____cpython:
+        CONFIG: linux_aarch64_c_compiler_version12cuda_compilerNonecuda_compiler_versionNonepython3.8.____cpython
         UPLOAD_PACKAGES: 'True'
         DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cos7-x86_64
-      linux_ppc64le_cuda_compilerNonecuda_compiler_versionNonepython3.9.____cpython:
-        CONFIG: linux_ppc64le_cuda_compilerNonecuda_compiler_versionNonepython3.9.____cpython
+      linux_aarch64_c_compiler_version12cuda_compilerNonecuda_compiler_versionNonepython3.9.____cpython:
+        CONFIG: linux_aarch64_c_compiler_version12cuda_compilerNonecuda_compiler_versionNonepython3.9.____cpython
         UPLOAD_PACKAGES: 'True'
         DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cos7-x86_64
-      linux_ppc64le_cuda_compilernvcccuda_compiler_version11.2python3.10.____cpython:
-        CONFIG: linux_ppc64le_cuda_compilernvcccuda_compiler_version11.2python3.10.____cpython
+      linux_ppc64le_c_compiler_version10cuda_compilernvcccuda_compiler_version11.2python3.10.____cpython:
+        CONFIG: linux_ppc64le_c_compiler_version10cuda_compilernvcccuda_compiler_version11.2python3.10.____cpython
         UPLOAD_PACKAGES: 'True'
         DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cuda:11.2
-      linux_ppc64le_cuda_compilernvcccuda_compiler_version11.2python3.11.____cpython:
-        CONFIG: linux_ppc64le_cuda_compilernvcccuda_compiler_version11.2python3.11.____cpython
+      linux_ppc64le_c_compiler_version10cuda_compilernvcccuda_compiler_version11.2python3.11.____cpython:
+        CONFIG: linux_ppc64le_c_compiler_version10cuda_compilernvcccuda_compiler_version11.2python3.11.____cpython
         UPLOAD_PACKAGES: 'True'
         DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cuda:11.2
-      linux_ppc64le_cuda_compilernvcccuda_compiler_version11.2python3.8.____cpython:
-        CONFIG: linux_ppc64le_cuda_compilernvcccuda_compiler_version11.2python3.8.____cpython
+      linux_ppc64le_c_compiler_version10cuda_compilernvcccuda_compiler_version11.2python3.8.____cpython:
+        CONFIG: linux_ppc64le_c_compiler_version10cuda_compilernvcccuda_compiler_version11.2python3.8.____cpython
         UPLOAD_PACKAGES: 'True'
         DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cuda:11.2
-      linux_ppc64le_cuda_compilernvcccuda_compiler_version11.2python3.9.____cpython:
-        CONFIG: linux_ppc64le_cuda_compilernvcccuda_compiler_version11.2python3.9.____cpython
+      linux_ppc64le_c_compiler_version10cuda_compilernvcccuda_compiler_version11.2python3.9.____cpython:
+        CONFIG: linux_ppc64le_c_compiler_version10cuda_compilernvcccuda_compiler_version11.2python3.9.____cpython
         UPLOAD_PACKAGES: 'True'
         DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cuda:11.2
-      linux_ppc64le_cuda_compilernvcccuda_compiler_version11.8python3.10.____cpython:
-        CONFIG: linux_ppc64le_cuda_compilernvcccuda_compiler_version11.8python3.10.____cpython
+      linux_ppc64le_c_compiler_version11cuda_compilernvcccuda_compiler_version11.8python3.10.____cpython:
+        CONFIG: linux_ppc64le_c_compiler_version11cuda_compilernvcccuda_compiler_version11.8python3.10.____cpython
         UPLOAD_PACKAGES: 'True'
         DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cuda:11.8
-      linux_ppc64le_cuda_compilernvcccuda_compiler_version11.8python3.11.____cpython:
-        CONFIG: linux_ppc64le_cuda_compilernvcccuda_compiler_version11.8python3.11.____cpython
+      linux_ppc64le_c_compiler_version11cuda_compilernvcccuda_compiler_version11.8python3.11.____cpython:
+        CONFIG: linux_ppc64le_c_compiler_version11cuda_compilernvcccuda_compiler_version11.8python3.11.____cpython
         UPLOAD_PACKAGES: 'True'
         DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cuda:11.8
-      linux_ppc64le_cuda_compilernvcccuda_compiler_version11.8python3.8.____cpython:
-        CONFIG: linux_ppc64le_cuda_compilernvcccuda_compiler_version11.8python3.8.____cpython
+      linux_ppc64le_c_compiler_version11cuda_compilernvcccuda_compiler_version11.8python3.8.____cpython:
+        CONFIG: linux_ppc64le_c_compiler_version11cuda_compilernvcccuda_compiler_version11.8python3.8.____cpython
         UPLOAD_PACKAGES: 'True'
         DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cuda:11.8
-      linux_ppc64le_cuda_compilernvcccuda_compiler_version11.8python3.9.____cpython:
-        CONFIG: linux_ppc64le_cuda_compilernvcccuda_compiler_version11.8python3.9.____cpython
+      linux_ppc64le_c_compiler_version11cuda_compilernvcccuda_compiler_version11.8python3.9.____cpython:
+        CONFIG: linux_ppc64le_c_compiler_version11cuda_compilernvcccuda_compiler_version11.8python3.9.____cpython
         UPLOAD_PACKAGES: 'True'
         DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cuda:11.8
+      linux_ppc64le_c_compiler_version12cuda_compilerNonecuda_compiler_versionNonepython3.10.____cpython:
+        CONFIG: linux_ppc64le_c_compiler_version12cuda_compilerNonecuda_compiler_versionNonepython3.10.____cpython
+        UPLOAD_PACKAGES: 'True'
+        DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cos7-x86_64
+      linux_ppc64le_c_compiler_version12cuda_compilerNonecuda_compiler_versionNonepython3.11.____cpython:
+        CONFIG: linux_ppc64le_c_compiler_version12cuda_compilerNonecuda_compiler_versionNonepython3.11.____cpython
+        UPLOAD_PACKAGES: 'True'
+        DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cos7-x86_64
+      linux_ppc64le_c_compiler_version12cuda_compilerNonecuda_compiler_versionNonepython3.8.____cpython:
+        CONFIG: linux_ppc64le_c_compiler_version12cuda_compilerNonecuda_compiler_versionNonepython3.8.____cpython
+        UPLOAD_PACKAGES: 'True'
+        DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cos7-x86_64
+      linux_ppc64le_c_compiler_version12cuda_compilerNonecuda_compiler_versionNonepython3.9.____cpython:
+        CONFIG: linux_ppc64le_c_compiler_version12cuda_compilerNonecuda_compiler_versionNonepython3.9.____cpython
+        UPLOAD_PACKAGES: 'True'
+        DOCKER_IMAGE: quay.io/condaforge/linux-anvil-cos7-x86_64
     maxParallel: 32
   timeoutInMinutes: 360