--- a/.scripts/run_osx_build.sh
+++ b/.scripts/run_osx_build.sh
@@ -26,10 +26,10 @@ conda activate base
 mamba install --update-specs --quiet --yes --channel conda-forge --strict-channel-priority \
     pip mamba conda-build boa conda-forge-ci-setup=3
 mamba update --update-specs --yes --quiet --channel conda-forge --strict-channel-priority \
-    pip mamba conda-build boa conda-forge-ci-setup
+    pip mamba conda-build boa conda-forge-ci-setup=3
 
 
-conda uninstall --quiet --yes --force conda-forge-ci-setup
+conda uninstall --quiet --yes --force conda-forge-ci-setup=3
 pip install --no-deps recipe/.
 
 echo -e "\n\nSetting up the condarc and mangling the compiler."