--- a/recipe/cross_compile_support.sh
+++ b/recipe/cross_compile_support.sh
@@ -98,7 +98,13 @@ if [[ "${HOST_PLATFORM}" != "${BUILD_PLATFORM}" ]]; then
                 sed 's/"//g' versions.txt | sed 's/_/-/g' | sed 's/-api//g' | sed 's/-dev/-devel/g' | sed 's/-develel/-devel/g' > rpms.txt
 
                 # filter packages from manifest down to what we need for cross-compilation
-                grep -E "cuda-(compat|cudart|cupti|driver|nvcc|nvml|nvprof|nvrtc|nvtx|profiler).*|lib(cu|npp|nvjpeg).*" rpms.txt > rpms_cc.txt
+                grep -E "cuda-(compat|cudart|cupti|driver|nvcc|nvml|nvprof|nvrtc|nvtx).*|lib(cu|npp|nvjpeg).*" rpms.txt > rpms_cc.txt
+
+                # the cuda-profiler-api package is not listed in the manifest, so we
+                # cannot generate it automatically
+                if [[ "${CUDA_COMPILER_VERSION}" == "11.8" ]]; then
+                    echo "cuda-profiler-api:11.8.86" >> rpms_cc.txt
+                fi
 
                 echo "Installing the following packages (<pkg>:<version>)"
                 cat rpms_cc.txt