https://github.com/conda-forge/magma-feedstock/pull/35#issuecomment-1745778381

Possibly we need to add the cuda profiler api package to CUDA 11.8 cross compiler environments, in order to fix

```
$SRC_DIR/sparse/blas/zgeisai_batched32.cu:16:10: fatal error: cuda_profiler_api.h: No such file or directory
```

When cross compiling for CUDA 11.8.

The `cuda-profiler-api` RHEL package only exists for 11-8 and 12-x, so I'm not sure what package name belongs in the "old" section.


<!--
Thank you for your pull request.
Below are a few things we ask you kindly to self-check before getting a review. Remove checks that are not relevant.
-->
Checklist
* [x] Used a [personal fork of the feedstock to propose changes](https://conda-forge.org/docs/maintainer/updating_pkgs.html#forking-and-pull-requests)
~~* [ ] Bumped the build number (if the version is unchanged)~~
* [x] Reset the build number to `0` (if the version changed)
* [x] [Re-rendered]( https://conda-forge.org/docs/maintainer/updating_pkgs.html#rerendering-feedstocks ) with the latest `conda-smithy` (Use the phrase <code>@<space/>conda-forge-admin, please rerender</code> in a comment in this PR for automated rerendering)
* [ ] Ensured the license file is being packaged.

<!--
Please note any issues this fixes using [closing keywords]( https://help.github.com/articles/closing-issues-using-keywords/ ):
-->

<!--
Please add any other relevant info below:
-->