@jakirkham 2023-10-06
Thanks Daniel! 🙏

Had one suggestion above