@carterbox 2023-10-06
I'm pretty sure I did the most sensible thing in order to add the profiler-api package. How do I test this? If I build the magma-feedstock locally in Docker, it will pull from the internet and not use my locally build packages?