@jakirkham 2023-10-06
Thanks Daniel! 🙏

Yeah would run this build locally. Then hardlink the `build_artifacts` from this feedstock directory into the feedstock directory to build (so they share the same `build_artifacts` directory). That way this package would be treated as part of the local package cache that can be installed