@carterbox 2023-10-07
Linking the build_artifacts directory does not work. Maybe because conda-forge-ci-setup needs to be installed outside (before) of the conda-build context?