@jakirkham 2023-10-09
Don't think so. All these steps would happen within the container as part of setup

May need to add `--use-local` to [these lines]( https://github.com/conda-forge/magma-feedstock/blob/9a8ec96be73a83d163978ed5b9bcdd18239b770b/.scripts/build_steps.sh#L35-L38 )

Perhaps it is worth pushing these testing changes to some PR so we can discuss any other changes that might be needed (though hoping `--use-local` is all we need)