@jakirkham 2023-10-10
Do the logs show which `conda-forge-ci-setup` package is pulled in now? If so, maybe we can figure out what variant is relevant based on that